Track if tokens have leading newline (ASI)
In order to implement ASI correctly, we need to know if there is a
newline between two tokens. Add a flag to indicate whether a token had a
newline between it and the previous token.

This commit doesn't make the parser use the flag yet. This commit should
not change behavior of existing code.