CI: flatten build matrix; prepare for Windows builds
I want to add Windows builds to GitHub Actions CI. Windows doesn't have
ASAN+UBSAN yet, and if it did, it wouldn't use the same flags as GCC and
Clang. Move the ASAN+UBSAN flags into the toolchain matrix section so,
in the future, CI doesn't try to build on Windows with ASAN+UBSAN.

This commit should not change behavior.