Use expression parser to parse let and parameter bindings
The parser for 'let' bindings and the parser for expressions have a lot
of overlap. Consolidate this duplicate functionality by making the 'let'
parser use the expression parser. Similarly, make the parsers for
function parameters and imports use the expression parser too.

This commit changes the behavior of a bunch of untested cases. The
chagned behavior will be refined when the parser is polished; I have no
plans to fix the introduced differences soon.