Refactor object literal expression access
In order to access the property-value pairs of an object literal, users
of the expression class must iterate children two at a time. This is a
bit fragile and isn't self-documenting. This design also makes it
difficult to expose flags for properties (such as whether a property is
a getter or setter).

Change users of object literal expressions to use a new
object_property_value_pair interface instead. This simplifies users and
enables future flexibility.

As a part of this change, also prevent users of object literal
expressions from inspecting children directly.

This commit should not change behavior.