Move error reporter into own module; add tests
debug_error_reporter in main.cpp is untested. I want to make
improvements to error reporting, but I'm scared of breaking things. Move
debug_error_reporter into its own module so that it can be tested, write
tests for it, and rename it to text_error_reporter.

This commit should not change behavior.