[2/2] Output JSON for Vim integration
Add a --output-format=vim-qflist-json option to the CLI. This option
makes quick-lint-js emit JSON which can be consumed by Vim (using Vim's
json_decode() function) and given directly to Vim's setqflist()
function for in-line diagnostics.

This commit should not change the default behavior. The old output
format (--output-format=gnu-like) is still the default.