Add --vim-file-bufnr option
Sometimes, the file name given to quick-lint-js (and thus the file name
quick-lint-js would report) does not match what Vim thinks the file name
is. This can happen if a temporary file is given to quick-lint-js to
lint unsaved changes, for example.

To make quick-lint-js' output work directly with Vim's setqflist()
function, allow the user to customize the "bufnr" key of the output.

If the --vim-file-bufnr option is not given, this commit should not
change behavior.