Vim: write JSON to stdout; separate from assertion messages
If quick-lint-js crashes with --output-format=vim-qflist-json, the
output is poorly formed JSON:

    {"qflist": [{"col": 6, "lnum": 3, "end_col": 5, "end_lnum": 3, "vcol": 0, "bufnr": 3, "filename": "SNIP", "text": "missing semicolon after expression"},
    {"col": 13, "lnum": 3, "end_col": 12, "end_lnum": 3, "vcol": 0, "bufnr": 3, "filename": "SNIP", "text": "missing semicolon after expression"}../src/include/quick-lint-js/parse.h:155: fatal: token not implemented in parse_and_visit_statement: _from on line 3 column 14

(Note the assertion failure message at the end of the second line.)

Separate the assertion message from the JSON by printing the latter on
stdout (instead of stderr). In the case of a crash, the JSON will still
be malformed. However, a carefully-written JSON parser could
automatically close fix the JSON and extract the diagnostics.