Rename variable in linter
variables_used_before_declaration is a confusing name because elements
may or may not be delcared in the future. A better name might be
variable_used_with_no_declaration_or_before_declaration, but let's just
drop the 'before declaration' signal and just call it variables_used.

This commit should not change behavior.