Remove variables from variables_used when they are declared
If a variable is used but hasn't been declared yet, we add it to
variables_used. When we see a declaration, we keep it in variables_used,
but we don't need to. Prune variables_used when visiting a declaration.
In the future I think this will let us clean up some other code because
it can assume that variables_used only refers to undeclared variables.

This commit should not change behavior.