Refactor erasing loop in visit_variable_declaration
The loop in linter::visit_variable_declaration is pretty messy. It's
hard to verify that the iterator variables are being used properly.

Use std::remove_if to remove variables instead of manually maintaining
an iterator. This means we have side effects in std::remove_if's
predicate, but I think that's better than what we have now.