Report use-before-declaration for shadowed variables
quick-lint-js should report an error for the following code, but
doesn't:

    let x = 69;          // line 1
    for (let _ of []) {  // line 2
      x;                 // line 3
      let x = 42;        // line 4
    }

The linter thinks that the use of 'x' on line 3 refers to the variable
declared on line 1. However, the use on line 3 actually refers to the
variable declared on line 4. Therefore, the linter should report a
use-before-declaration error on line 3, but it doesn't.

Have the linter not look up variables in parent scopes until all
variable declarations for the current scope have been seen.