Add extra null bytes at end of source string
I want to add some SIMD magic to the lexer. To make this possible, we
must be able to read bytes outside the bounds of the string. (Otherwise,
we'd need to check if we're about to reach EOF before reading into SIMD
registers, which is undesirable.)

Create a padded_string container which enforces extra null bytes at the
end of the source string. Make the lexer use it.

This commit does not introduce any SIMD magic. It only allows future
SIMD magic.

This commit does not change behavior.