Fix wrong comparison in lexer
On platforms where char is signed, lexer::skip_whitespace has a
performance bug:

c >= static_cast<char>(0xc2) is interpreted as c >= -0x3e. This is
correct (it passes all of our tests), but not what was intended at all.
This check exists because non-ASCII codes are unlikely. Instead, for
ASCII-only source code, the switch is entered 100% of the time!

Fix the condition so we compare in unsigned land as intended. This
doesn't affect correctness, but it does improve reduce branching into
the switch.

Before:

    4968d: 3c c2              cmp $0xc2,%al
    4968f: 0f 8d fb 04 00 00  jge 49b90 <quick_lint_js::lexer::parse_current_token()+0x530>

After:

    4968d: 3c c1              cmp $0xc1,%al
    4968f: 0f 87 fb 04 00 00  ja  49b90 <quick_lint_js::lexer::parse_current_token()+0x530>