@@@ Optimize parse_identifier with digit lookup table
@@@ benchmark

In parse_identifier, we perform several operations to match digits and
undescore:

    >= '0'
    <= '9'
    &
    == '_'
    |

Optimize these operations using a table lookup approach. The lookup
relies on SSSE3's _mm_shuffle_epi8 intrinsic to use a vector as indexes
into a different vector. With _mm_shuffle_epi8(A, B), the bottom 4 bits
of each byte of B are an index into A. This lets us build a look-up
table as long as the bottom 4 bits of the bytes to look up are unique.
The look-up table approach has fewer operations:

    _mm_shuffle_epi8([0-9_], ...)
    ==

Implement this algorithm if the compiler and processor support SSSE3.

This commit should not change behavior.