Auto-declare 'foo' if we see 'typeof foo'
The typeof operator has special semantics. typeof is often used to see
if a variable exists in the global scope. Make quick-lint-js assume that
'typeof foo' means that a variable 'foo' exists in the global scope (if
'foo' isn't otherwise declared).

With this commit, jquery-3.5.1.js parses and lints with no errors.
Hurray!