Fix crash when using hoisted vars
In the following code, the declaration of x is hoisted out of the block
scope, so the use of x binds to the declared variable:

    x;
    {
      var x;
    }

This code lints fine if it's in global scope, but crashes quick-lint-js
if it's in function scope.
propagate_variable_declarations_to_parent_scope propagates the
declaration, but doesn't remove the use of x from scope::variable_uses.
This left-over x causes the following assertion in
propagate_variable_uses_to_parent_scope to fail:

    for (const used_variable &used_var : current_scope.variables_used) {
      QLJS_ASSERT(!current_scope.find_declared_variable(used_var.name));

Fix the crash by making propagate_variable_declarations_to_parent_scope
do the same work that visit_variable_declaration does.