WASM: add Node.js CLI
I want to use quick-lint-js in a web browser. Add the glue to make it
possible to call quick-lint-js code from JavaScript code in Node.js.

To run the code included in this commit, install the emscripten SDK [1]
then run the following commands:

    $ emcmake cmake -S . -B build-emscripten -G Ninja
    $ emmake ninja -v -C build-emscripten quick-lint-js-wasm-demo
    $ node demo/test.mjs
    $ node demo/cli.mjs hello.js

Future commits will integrate quick-lint-js with the browser.

[1] https://emscripten.org/docs/getting_started/downloads.html