WASM: add demo web page
I want to demonstrate quick-lint-js' abilities. Make a web page with a
text editor which highlights errors inline.

The demo page has a lot of missing features and a lot of bugs. Some day,
this demo page will be improved.

In order to run the demo properly, you must first build quick-lint-js
with emscripten into a directory named 'build-emscripten'. The workflow
will be improved in the future.