Don't interpret \u{69}f as a keyword
Identifiers containing Unicode escape sequences should not ever be
parsed as keywords. Teach this rule to quick-lint-js.

Additionally, if an identifier would have the same name as a keyword,
that's a syntax error according to the standard. Report an error in such
cases.

Contextual keywords and property names are still a bummer, but we'll
handle that later.