Add nix build expressions and instructions
This adds a new directory containing packaging tools dist/. For now,
this only includes nix derivations which allow building and packaging
through the nix package manager[1].

Most of the work for the nix derivation is done by the build hook for
cmake[2]. We include a default.nix and shell.nix for convenience.
While default.nix sets buildType Release, shell.nix sets buildType
to Debug. This does the right thing most of the time.

We add a new file INSTALLING.md which describes installation procedure
with nix-env, and also extend BUILDING.md to explain how to use nix
for development.

[1] https://nixos.org/manual/nixpkgs/stable/
[2] https://nixos.wiki/wiki/C#cmake