Error if certain variables are named 'let'
Variables declared with 'var' can be called 'let', but variables
declared with 'const' cannot. Report an error in the latter case, and in
a few related latter cases.

Unfortunately, the current implementation does a string comparison for
each declared variable. I haven't measured, but this could negatively
impact performance if a variable is not named 'let'.