Avoid string comparison for 'let'
When lexing, we already know whether an identifier is "let" or not.
The parser knows this too, but discards this information when building
expression ASTs. Preserve this information in the AST to avoid a string
comparison for let/const/import variables.

I ran the parser benchmarks on my machine with jQuery (worst case for
this commit, because jQuery never uses let or const or import). I
observed a 0.2% slowdown:

* I suspect the slowdown is caused by storing to
  expression::variable::type_. These stores are not needed by the jQuery
  benchmark.
* The storage of expression::variable::type_ costs no space. We slipped
  type_ in the padding bits already present before variable_identifier_.

This commit should not change behavior.