Parse for(var v = initializer of []) {}
for-of loops can contain a 'var' variable declaration with an
initializer. (for-in loops cannot, and for-of loops with 'let' or
'const' declarations cannot.) Parse such weird code.

This commit does change the order of visits for some code. This should
be fine for now, because 'var' variables are hoisted. It might cause
problems for future lint rules, however.