Optimize read_file buffer size (POSIX)
stat::st_blksize is the recommended size of buffers given to read(). Use
this size in our read_file function.

I didn't measure the performance impact of this change. It probably
doesn't matter one bit for realistic workloads. (Typically, we'll
successfully get the file's size, so we don't need to read in chunks.)

This commit should not change behavior.