Move wasm demo code into demo/ directory
A wildcard CLI build (e.g. `g++ src/*.cpp ...`) builds the bindings for
the WASM demo. This is a mistake which might cause problems in the
future.

Move the WASM bindings into the demo/ directory so wildcard builds don't
over-build. This also declutters src/CMakeLists.txt.

This commit should not change behavior.