Make posix_fd_file::duplicate less awkward
posix_fd_file::duplicate sticks out because it's a static function and
is used with a non-owning file descriptor in tests. Create a separate
class, posix_fd_file_ref, with a non-static duplicate function. This
class is less awkward in my opinion, and it will be a nice place to put
other functions needing non-owning file descriptors.

This commit should not change behavior.