Implement int-to-string conversion
For JSON writing in the upcoming Language Server Protocol server, we
need an integer-to-string conversion routine. In other areas of
quick-lint-js, we use std::ostream's conversion. I want to reduce our
reliance on std::ostream, so we need to make the integer-to-string
function ourselves. Implement it using std::to_chars if available, or
std::snprintf otherwise.

This commit should not change behavior. The implemented write_integer
function is not yet used anywhere.