Make global scope different type than other scopes
In the linter class, the global scope has the same representation as all
other scopes. However, the global scope has different properties, so I
want it to have a different representation.

Refactor the linter so the global scope has its own type. Keep the
representation mostly the same for now, though.

This commit should not change behavior.