Parse exported variables as variable uses
When exporting variables as in the following code, the variables
mentioned are considered referenced/used:

    export {taco, banana};

Currently, the parser ignores these uses. Tell the parser to emit
visit_variable_export_use for such uses.

This commit only teaches the parser about these variable uses. This
commit does not teach the linter about these variable uses, because the
rules are not the same as normal variable uses, and I want this commit
to be somewhat readable.