Speed up compilation of error_collector
Profiling shows that compiling the error_collector::error overloads is
expensive. I suspect this is due to creating all the overloads of
vector<>::emplace_back and creating all the overloads of
error_variant::error_variant.

On my machine, this commit reduces compile times from 4.390 seconds to
3.166 seconds.

This commit should not change behavior.