Fix compilation error on newer MSVC
We are calling GetNamedPipeInfo incorrectly. Its second parameter is a
pointer, but we are giving an integer.

I think the code happened to compile with older versions of MSVC because
the integer we give constant-folds to 0, and 0 is apparently coerced
into a null pointer. Newer versions of MSVC Do The Right Thing and don't
treat the computed 0 as a null pointer, leading to a compilation error.

Fix the call to GetNamedPipeInfo, providing a null pointer as desired.