LSP: Refactor: Remember document text
LSP has a mechanism for incremental document changes. This mechanism
requires the server (quick-lint-js) to keep track of the text of
documents. Refactor linting_lsp_server_handler so it remembers the
text of loaded documents in addition to their URIs.

This commit should not change behavior. This commit does not implement
incremental document changes.