Refactor parse_and_visit_let_bindings
In JavaScript, loop labels to be called 'let'. quick-lint-js' parser
doesn't support this. The skipping of the 'let' token is handled in
parse_and_visit_let_bindings, making it difficult to parse loop labels
properly.

Make parse_and_visit_let_bindings accept the 'let' token as a parameter
instead of having it peek at the current token.

This commit should not change behavior.