Disallow illegal non-ASCII escaped characters in identifiers
Character escape sequences in identifiers can refer to code points which
are illegal for identifiers. quick-lint-js reports errors for such
illegal code points if they are ASCII, but does not report illegal
non-ASCII code points.

Generate a table of code point identifier legality, and consult the
table to reject characters per the standard.