Use char32_t for code points
In lexer, we were mixing int and char32_t for code points. Use char32_t
consistently.

This change brought to light some sketchy uses of
is_identifier_character. is_identifier_character takes a Unicode code
point, but some callers are giving a UTF-8 code unit instead. This is
wrong. We'll fix that in the future.

Hopefully, this commit should not change behavior, but I wouldn't bet on
it.