Fix ASCII optimization when inserting text with newlines
In several scenarios, lsp_locator's is-line-ASCII cache gets out of sync
when inserting text containing a newline character. This causes
characters to be counted incorrectly. Fix this bug.

I think the current implementation is still incorrect; it might think an
ASCII line is non-ASCII. This is only a performance problem, not a
correctness problem.