Create tool for generating copyright documentation
I want to distribute binaries of quick-lint-js. In order to do this, we
must comply with licenses of third-party libraries.

Write a tool which inspects ELF linkmaps to generate documentation which
(hopefully) complies with licenses of third-party software.

I tested the tool on my Linux Mint (Ubuntu/Debian/dpkg-based) machine
with a statically-linked quick-lint-js executable.