CI: Make testing builds easier
We only save build artifacts to long-term storage for master builds.
This makes it hard to test things. Save builds for any branch, not just
builds for master.

We still don't save builds for PRs, because that is a possible security
risk. (A workflow change in a PR could upload a PHP script which the
web server might execute, for example.)