[2/3] Give codes to errors: CLI, Vim, LSP
Error messages might change, be translated, or have generated components
within them. This makes messages difficult to google.

When reporting errors via the CLI, Vim, or LSP, include an error code
with the message. This error code should be easier to google (if
quick-lint-js ever becomes popular).

A future commit will add the code to the Visual Studio Code plugin.