Make write_fatal_error_* always write to stderr
write_fatal_error_unimplemented_token and
write_fatal_error_unimplemented_character write messages for debugging
purposes. Therefore, it makes little sense for these functions to write
to a std::ostringstream in tests, for example.

Force these functions to write to stderr.

In practice, for users of the CLI or LSP server or VS Code plugin, this
commit should not change behavior.