Refactor async arrow function parameter parsing
When the parser sees an 'async' token followed by a '(' token, it reads
an expression starting at the '(', then unpacks the expression into the
parameters (with a special case for '()'). This makes it a bit harder to
make some nice diagnostics for certain errors, such as extraneous commas.

Refactor the code into a simple loop, similar to how we parse function
call arguments.

This commit should not change behavior.