Error on 'for (async of []);'
For a while, the code 'for (async of []);' is ambiguous in the JS
standard. It's going to be made unambiguously a SyntaxError. Make
quick-lint-js report an error for this code, treating 'async' as a
variable name, and gracefully continue.

Many other forms, such as 'for (async.prop of []);' and
'for (async of => {}; cond; update);' are unambiguous, and shouldn't
result in any errors. This commit adds tests for these similar cases.

Discussion on the ambiguity and its resolution:
https://github.com/tc39/ecma262/issues/2034