Implement lexer transactions
In several spots in the parser, we want to look ahead a few tokens in
order to implement different behavior. For example, when parsing
'for (async of', we need to see the token after 'of' to determine
whether to parse 'async' as a keyword (for an async arrow function) or
as an expression (for a for-of loop).

The parser implements lookahead with a lot of code repetition. A better
approach might be to implement backtracking.

Make backtracking possible by adding a transactions API to the lexer.
The transactions API lets the parser undo calls to lexer::skip.

The transaction feature is implemented by saving and restoring the lexer
state. This means that, after rollback, tokens need to be re-parsed. I
expect that rollbacks will be rare, so re-parsing should be fine. A
different approach might be necessary if rollbacks are common.

This commit should not change behavior. This commit does not use the new
API, only implements it. A future commit will experiment with the new
API.