Improve bad token recovery in variable declarations
In a 'let' statement, if we see an unexpected token where we expect a
variable name, we currently recover by deleting the token and ending the
'let' statement. This often leads to more parse errors.

Change the error recovery such that the offending token is not deleted.
Instead, end the 'let' statement and parse the offending token as the
beginning of a new statement.