Disallow variables named 'enum', 'instanceof', etc.
'enum', 'instanceof', etc. are a reserved words, thus cannot be the name
of a variable. Disallow them.

I switch the tests from using 'let' to using 'var' because
'let instanceof C' is an expression statement, not a variable
declaration, so parse_and_visit_let_binding isn't called for
'let instanceof'.