Test \u{55}nicode escapes in keywords with parser
Currently, the lexer handles Unicode escape sequences in identifiers and
keywords. This is problematic, because whether an escape sequence is
legal depends on the context. For example:

    let o = { if: 42 };
    console.log(o.if);
    console.log(o.\u{69}f);  // OK
    \u{69}f () {}            // INVALID

We need the parser to decide whether sequences are allowed or not. As
such, the tests belong in the parser.

Copy the lexer tests into the parser.