Move keywords-cannot-contain-escape-sequences checking into parser
Whether an escape sequence is allowed in a keyword or not depends on the
parser's context. Move the decision of whether to report an error for
such keywords/identifiers into the parser.

This commit causes QLJS_PARSER_UNIMPLEMENTED to be called for more
inputs. Future commits will patch these holes.