Refactor boolean parameter in parser
parse_and_visit_named_exports takes a boolean parameter, which is an API
design smell in my opinion. Also, in the future, I want the signature of
exports to be different from the signature for imports.

Split the function's interface into two different functions, making the
calling code clearer and enabling future features.

This commit should not change behavior.