Refactor allocation of identifier escape sequence list
Currently, the token class is very lightweight and is copied around by
various parts of the parser under the assumption that copying is cheap.

I want to store the list of identifier escape sequences in the token
class. Currently, the list is represented as a std::vector<>. Adding a
std::vector<> member to the token class would introduce a lot of code
bloat and might negatively impact performance throughout the parser.

Make referencing the identifier escape sequences cheap:

1. Allocate the vector's contents on the lexer's monotonic allocator.
2. Allocate the vector itself on the lexer's monotonic allocator.
3. Pass around pointers to the vector instead of passing around the
   vector.

A future commit will modify the token class. This commit only changes
lexer's internal representation for now.

This commit should not change behavior.