Fix crash on non-ASCII whitespace after keyword
When parsing an identifier or a keyword, the lexer enters a slow path if
it sees a non-ASCII character, including non-ASCII whitespace. By
entering the slow path, lexer::parse_identifier returns a non-null
escape_sequences vector, even if no escape sequences were parsed.

This present-but-empty escape_sequences vector tricks the identifier
parsing code in lexer::parse_current_token into treating the
identifier/keyword as if it has an escape sequence (turning the token
into an identifier or a reserved_keyword_with_escape_sequence). This
interpretation is wrong, and can cause crashes if
token::identifier_escape_sequences is assumed to be non-empty.

Fix the crash by treating a present-but-empty escape_sequences vector
the same as a missing escape_sequences vector.