Add --exit-fail-on option to CLI
I want check-test262.go to be able to check for only syntax errors. To
make this easy, add a --exit-fail-on option to the CLI to allow users to
configure which errors are considered errors for the program's exit
status code.

If --exit-fail-on is unspecified, this commit should not change
behavior.