Remove tests if the platform doesn't support them
If the platform doesn't support some OS features used in a test (such as
emscripten not supporting googletest death tests), we replace the test
with a DISABLED_ test which fails. This is confusing, because running
the disabled test will not attempt to run the intended test code.

Instead of making a DISABLED_ placeholder test for uncompilable tests,
just hide the test entirely.