test262: Refactor test todo logic
When looking for files to test, we check each file against a blacklist
("todo"). This checking gives a yes-or-no answer. In the future, I want
to collect more information when checking each file, such as whether a
quick-lint-js failure is expected or unexpected.

Refactor todo checking to give a struct with possibly more information
than just yes or no.

This commit should not change behavior.