test262: Check that quick-lint-js errors if error is expected
If a test262 test says that it should fail with an early error, make
sure quick-lint-js reports an error for that test.

The current technique does not check which kind of error is emitted by
quick-lint-js, only that any error is emitted. Checking the error codes
will be done in the future. (quick-lint-js reports an false positive for
many tests because of: use of undeclared variable: $DONOTEVALUATE
[E057])

The current technique does not check that quick-lint-js reports no
errors for tests which should parse without error. This feature will be
done in the future.