Don't confuse variables from descendant scopes
In the following code, quick-lint-js reports that the assignment to 'b'
is illegal because it is assigning to a 'const' variable:

    const a = () => {
      b = null;
    };
    let b;

when processing the declaration of 'a', quick-lint-js accidentally
associates the assigment to 'b' with the variable 'a'. This is caused by
a missing check on the name. Very obvious in hindsight.

Fix the bug by checking the name when comparing variables.