Fix error when assigning to undeclared variable
If you assign to an undeclared variable in global scope, quick-lint-js
correctly reports an assignment error. However, if you assign to an
undeclared variable inside a function, quick-lint-js reports the issue
as a variable use error.

Improve the error so quick-lint-js reports the same kind of error
regardless of where the assignment is.