Fix tools/format on non-Linux
If nproc isn't installed, running nproc fails, and because of 'set -e',
the failure causes the tools/format script to exit. This is
unintentional; we want the script to continue even if nproc fails.

Fix tools/format to work on macOS and other systems where nproc might
not be installed.