Fix crash on broken shebang
Some test262 tests have something similar to the following code:

    #\u{21} // totally not a shebang

quick-lint-js now crashes on these tests because it treats the #\u{21}
as a private_identifier token (whereas previously the # would be skipped
and the \u{21} would be treated as an identifier token).

Fix the crash by complaining about and skipping the private_identifier.