Fix parsing of non-ASCII in RegExp literal flags
U+05d0 encoded in UTF-8 is 0xd7 0x90. U+00d7 is not a legal character in
identifiers.

Flags for regular expression literals are IdentifierPart according to
the standard. When we parse these flags, we check if the first byte
(char8_t) is a legal *character* (char32_t). (We interpret a byte as if
it was a Unicode code point.) This is wrong for U+05d0, whose first byte
is 0xd7 (char8_t). This causes the RegExp literal to be parsed
incorrectly.

Fix this issue by not mixing UTF-8 bytes with Unicode code points.