Fix leaks on QLJS_PARSER_UNIMPLEMENTED
In the LSP server, QLJS_PARSER_UNIMPLEMENTED calls longjmp to jump back
to the beginning of parsing. This means that stack-allocated variables
won't have their destructor called. This includes vectors, which might
have allocated memory on the heap. Because the destructors of vectors
aren't called, memory is leaked.

Fix the memory leak by allocating memory for vector data in a heap local
to the parser object. monotonic_buffer_resource might perform worse or
use more memory compared to the default/system allocator, but I think
it's worth avoiding memory leaks in the LSP server.