Website: Synchronize demo's pre and textarea sizing
https://user-images.githubusercontent.com/32761424/115054415-2f4cea00-9eae-11eb-9034-0e7e471d6f07.mp4





https://user-images.githubusercontent.com/32761424/115054442-34aa3480-9eae-11eb-9bc6-a98fba9f0a50.mp4





There seems to be some misalignment if you make the height too small. (potential "bug" not worth caring about)

I wasn't sure if the comment from the known bugs section referred to the fix I've implemented so I gave it it's own commit.
If it is referring to the fix, that commit should be fixup'd. If not, that commit should be dropped.