LSP: Prevent deadlock with synchronous client
If an LSP client sends a bunch of requests, but does not read from
quick-lint-js' stdout, quick-lint-js can hang in pipe_reader::write.
(This came up when writing some benchmarks. My LSP client naively made a
bunch of requests without reading responses during the requests.)

Prevent this (unlikely) deadlock by calling Writefile/writev/write for
stdout on a separate thread.

In a future commit, on POSIX platforms, we will switch stdout to
non-blocking mode and use an event loop (poll). Much of the code
necessary for this future solution (non_blocking_pipe_writer) is
included but disabled in this commit.

Unfortunately, libstdc++ had an ABI break between Ubuntu 16.04 and
Ubuntu 20.04 for std::thread. This causes the dpkg installation to fail
on Ubuntu 16.04:

    $ dpkg -i quick-lint-js_*.deb
    Selecting previously unselected package quick-lint-js.
    (Reading database ... 4785 files and directories currently installed.)
    Preparing to unpack quick-lint-js_0.2.0-1_amd64.deb ...
    Unpacking quick-lint-js (0.2.0-1) ...
    dpkg: dependency problems prevent configuration of quick-lint-js:
     quick-lint-js depends on libstdc++6 (>= 6); however:
      Version of libstdc++6:amd64 on system is 5.4.0-6ubuntu1~16.04.12.

    dpkg: error processing package quick-lint-js (--install):
     dependency problems - leaving unconfigured
    Errors were encountered while processing:
     quick-lint-js

Disable testing the Debian package on Ubuntu 16.04. When we set
QLJS_PIPE_WRITER_SEPARATE_THREAD to 0 in the future, we can re-enable
testing on Ubuntu 16.04.