Avoid redundant copying in lsp_pipe_writer on Windows
On Windows, lsp_pipe_writer needs to put the data to send into a single
continuous array. This is done by first prepending the header to the
byte_buffer then copying the byte_buffer's data into an array.

The prepend involves an allocation and a copy. This is a bit wasteful.

Reduce copying and allocations by writing the header into the array
directly instead of into the byte_buffer.

I didn't attempt to measure any performance improvement from this
change.

This commit should not change behavior.