Rename variables in lsp_pipe_writer to reduce confusion
lsp_pipe_writer's code uses the name 'message' for two different kinds
of things: a JSON-RPC message, and a JSON-RPC message with an LSP
header. Change the names such that 'message' (mostly) refers to a
JSON-RPC message, and 'data' refers to a message with an LSP header.

This commit should not change behavior.