Parse 'await /re/' and 'await / n' in top level
Currently, top-level await is implemented by checking the token
following the 'await' keyword. If the token looks like an operator,
'await' is treated as an identifier. Otherwise, 'await' is treated as an
operator. This technique doesn't work for '/', however, because '/'
could start a regular expression or be a binary operator.

Perform speculative parsing to determine whether 'await' followed by a
'/' should be interpreted as an identifier or as a unary operator. This
speculative parsing is probably very slow, but 'await/' should be super
rare in real-world programs anyway, so I'm not worried about potential
performance problems.