Website: Move .html files around
The .html files are poorly organized. Some files are in the root, and
some are in their own directory.

Organize the .html files such that each .html file gets its own
directory (except for the root index.html). Provide redirects so old
references to /hiring.html (for example) will automatically redirect the
user to the new URL.