Website: generate HTML chart+table with EJS
The generate-benchmark-html Python script creates an .html file based on
a .template.html file. The generated .html file is checked into the
repository, meaning that it needs to be manually synchronized with the
source .json benchmark data.

Rewrite the generate-benchmark-html Python script in JavaScript and use
the generation code from an EJS template. The website build process
(`yarn build`) will generate an index.html file from the .json benchmark
data and the index.ejs.html template.

The generated code for /benchmarks/index.html should be functionally
unchanged by this commit.