CI: prevent mismatch between workflows and checkout
As of commit cadbfbb1db664e1016a76e7cc9b2ffcc55243e6f, GitHub Actions CI
does the following on pull requests:

1. Merge the pull request source with the destination
2. If the merge succeeded:
   a. Read workflow files from the merged commit
   b. Run each workflow:
      i. 'git clone', creating a checkout at the pull request source

Importantly, the Git commit used for the workflow files is different
from the Git commit used for the checkout. This causes problems with a
workflow file assumes a newer checkout. (For example, commit
52ed8937c74ca839f317e6a4eeb3679eae73aaaa changed a workflow file and the
checkout such that the workflow fails without an updated checkout.)

I don't know of a way to coerce GitHub to use the workflow files of the
pull request source.

Change our workflows to use GitHub's defaults: read workflow files from
the merged commit, and create a checkout from the merged commit. This
means that CI won't test the code seen in the pull request, but I think
that behavior is slightly better in the common case than the current
behavior.