Website: generate error documentation from .md files
Our quick-lint-js-generate-error-documentation program, written in C++,
reads .md files in docs/errors/, checks them for mistakes, then creates
an .html file for the website. The .html file is checked into the
repository.

Checked-in generated code is difficult to maintain. Rewrite
quick-lint-js-generate-error-documentation in JavaScript, integrating
the .html generation with our website build pipeline. This means the
.html file no longer needs to be checked into the repository.