Rename 'vscode' module to 'c-api'
I want to generalize the qljs_vscode_* and <quick-lint-js/vscode.h> APIs
to be usable in more contexts. (For example, it'd be nice if the web
demo used the same code.)

Rename vscode.h to c-api.h, and rename qljs_vscode_* to qljs_*.

qljs_vscode_diagnostic keeps its name because the location information
it conveys is specific to what VS Code expects. Other consumers of the
API might want different formats.

This commit should not change behavior.