CI: build website on PRs
The deploy-website.yml workflow packages the website. If this process
breaks, we don't know until we want to deploy the website.

Add a CI workflow which does most of the work of deploy-website.yml.

Keen readers will observe that build-website.yml does a lot more work
than deploy-website.yml. This is because deploy-website.yml is currently
broken. I'll fix it in a future commit.