Suggest `async` when `await` is followed by arrow function.
Attempts to fix issue #279.

Excluded the cases where `await` is followed by an `async` then
followed by an arrow function. In this case `await` must be an unary
operator.

This commit does not solve issue #278 just yet, but should be a
start. Here the error is only caught for top-level await and async
function, i.e cases where async is *not* treated as an identifier.