Attempts to fix #278.
It's done by checking if there is a arrow function after `await` even
if its treated as an identifier.

There is some code duplication between this commit and the earlier
one so I will try to work out a way to combine these two.

Also, I don't know enough JS to know if the tests are strict
enough to catch all the cases.