Improved code from last commit
Included function to check if the expression is side-effect-free,
although I have to check more carefully if the cases are done
correctly.

The cases that explicitly not side-effect-free are: dot, index,
function and named_functions.

From testing I found that arrow functions invalidate placing an `if`,
even if their children are side-effect-free. This case is treated
seperately. Maybe there are more cases where this happens and I must
think it through.