Install Vim plugin on Arch Linux and Homebrew
When a user installs quick-lint-js on Arch Linux via its package manager
or on macOS via Homebrew, the Vim plugin is installed incorrectly and
doesn't work out of the box. Change the installation location, making
the Vim plugin work out of the box (assuming ALE is installed) on Arch
Linux and with Homebrew.

This commit tries to preserve how the Vim plugin was installed on
Debian.