Website: squiggle under errors in error docs
Documentation for each error includes a code example demonstrating the
error. Highlight the error with a squiggly, just like how the web demo
("Try it online" page) highlights errors.

The implementation reuses the front-end code on the server (using
JSDOM), so it's a bit ugly and slow. But the code reuse is worth it for
now.